Remove useless components
