--- packages/core/admin/admin/src/components/Roles/Chevron/index.js
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-
-const Chevron = styled(FontAwesomeIcon)`
-  width: 18px !important;
-  display: none;
-  padding-left: 1rem;
-  font-size: 2.1rem;
-`;
-
-export default Chevron;
