--- packages/core/admin/admin/src/components/Roles/RoleForm/NameInput.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import SizedInput from '../../SizedInput';
-
-const NameInput = inputProps => <SizedInput name="name" type="text" {...inputProps} />;
-
-export default NameInput;
