--- packages/core/admin/admin/src/components/Roles/RoleList/index.js
@@ -1,2 +1,0 @@
-export { default as RoleListWrapper } from './RoleListWrapper';
-export { default as RoleRow } from './RoleRow';
