+++ packages/core/admin/admin/src/components/Roles/index.js
@@ -1,6 +1,4 @@
-export { default as ButtonWithNumber } from './ButtonWithNumber';
 export { default as EmptyRole } from './EmptyRole';
-export { default as Tabs } from './Tabs';
 export { default as Permissions } from './Permissions';
-export * from './RoleForm';
-export * from './RoleList';
+export { default as RoleForm } from './RoleForm';
+export { default as RoleRow } from './RoleRow';
